# aidis-command/frontend/.openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AIDIS Command Backend API",
    "version": "1.0.0",
    "description": "RESTful API for AIDIS Command Backend with MCP Bridge Integration",
    "contact": {
      "name": "AIDIS Development Team",
      "email": "dev@aidis.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3002/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiSuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "Response data payload"
          },
          "correlationId": {
            "type": "string",
            "format": "uuid",
            "description": "Request correlation ID for tracking"
          },
          "metadata": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "version": {
                "type": "string",
                "example": "1.0.0"
              }
            }
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "validation",
                  "authentication",
                  "authorization",
                  "not_found",
                  "internal",
                  "business"
                ],
                "description": "Error type for client handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error details (e.g., validation errors)"
              },
              "code": {
                "type": "string",
                "description": "Machine-readable error code"
              }
            },
            "required": [
              "type",
              "message"
            ]
          },
          "correlationId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
      "BaseEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at"
        ]
      },
...
      "ContextEntity": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseEntity"
          },
          {
            "type": "object",
            "properties": {
              "content": {
                "type": "string",
                "description": "Context body content"
              },
              "type": {
                "type": "string",
                "enum": [
                  "code",
                  "decision",
                  "research",
                  "issue",
                  "note",
                  "error",
                  "test"
                ],
                "description": "Context classification"
              },
              "tags": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Associated tags"
              },
              "metadata": {
                "type": "object",
                "additionalProperties": true,
                "description": "Arbitrary metadata payload"
              },
              "session_id": {
                "type": "string",
                "format": "uuid",
                "description": "Linked session identifier"
              },
              "project_id": {
                "type": "string",
                "format": "uuid",
                "description": "Associated project identifier"
              },
              "relevance_score": {
                "type": "number",
                "description": "Semantic similarity score when applicable"
              }
            },
            "required": [
              "content",
              "type"
            ]
          }
        ]
      },
      "UpdateContext": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Updated context content"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Updated tag list"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Updated metadata payload"
          },
          "relevance_score": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "description": "Updated relevance score"
          }
        }
      },
      "ContextSearchResponse": {
        "type": "object",
        "properties": {
          "contexts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContextEntity"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        },
        "required": [
          "contexts",
          "total"
        ]
      },
...
      "ContextBulkDelete": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "minItems": 1
          }
        },
        "required": [
          "ids"
        ]
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/contexts": {
      "get": {
        "summary": "Search contexts with filters",
        "tags": [
          "Contexts"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string"
            },
            "description": "Free-text search query"
          },
          {
            "in": "query",
            "name": "session_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "code",
                "decision",
                "research",
                "issue",
                "note",
                "error",
                "test"
              ]
            }
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated tags filter"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contexts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ContextSearchResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/contexts/search": {
      "post": {
        "summary": "Perform semantic context search",
        "tags": [
          "Contexts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContext"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Semantic search results",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ContextSearchResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/contexts/{id}": {
      "get": {
        "summary": "Get a context by ID",
        "tags": [
          "Contexts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Context retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ContextEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Context not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an existing context",
        "tags": [
          "Contexts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Context updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ContextEntity"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Context not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a context",
        "tags": [
          "Contexts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Context deleted successfully"
          },
          "404": {
            "description": "Context not found"
          }
        }
      }
    },
    "/contexts/bulk/delete": {
      "delete": {
        "summary": "Bulk delete contexts by ID",
        "tags": [
          "Contexts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextBulkDelete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contexts deleted successfully"
          }
        }
      }
    },
    "/contexts/{id}/related": {
      "get": {
        "summary": "Get contexts related to the supplied context",
        "tags": [
          "Contexts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Related contexts returned",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/ContextEntity"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}